QwenimageTurbo/requirements.txt
diffusers[torch]
transformers
accelerate
gradio
sentencepiece
protobuf
bitsandbytes
psutil
einops
torchao
scipy
scikit-image
image-reward
lpips
fastapi
uvicorn